Extract formatCleanResult to clean-formatter.mts to stay under 200-line limit
