+++ src/cli/formatters.mts
@@ -1,5 +1,6 @@
 export { formatIterateResult } from "./iterate-formatter.mts";
 export { projectIterateLean, projectIterateVerbose } from "./iterate-lean.mts";
+export { formatCleanResult } from "./clean-formatter.mts";
 
 import { safeFence } from "./fence.mts";
 import {
@@ -13,7 +14,6 @@ import { joinSections } from "../util/markdown.mts";
 import type { FetchResult } from "../commands/resolve.mts";
 import type { CommitSuggestionResult } from "../types.mts";
 import type { ResolveResult } from "../comments/resolve.mts";
-import type { CleanResult } from "../commands/clean.mts";
 
 export function formatFetchResult(result: FetchResult): string {
   const activeTotal =
@@ -139,30 +139,6 @@ export function formatCommitSuggestionResult(result: CommitSuggestionResult): st
     });
   }
   return lines.join("\n");
-}
-
-export function formatCleanResult(result: CleanResult): string {
-  if (!result.ok) {
-    return `Error: ${result.error ?? "unknown error"}`;
-  }
-
-  const heading = result.dryRun ? "## Would clean" : "## Cleaned";
-  const paths = result.dryRun ? result.deleted : result.deleted;
-
-  if (paths.length === 0) {
-    // Target did not exist.
-    const label = result.dryRun ? "Nothing to clean (dry-run)" : "Nothing to clean";
-    return `${label} — ${result.target} does not exist.`;
-  }
-
-  const sections: (string | null)[] = [
-    heading,
-    paths.map((p) => `- ${p}`).join("\n"),
-    result.dryRun
-      ? `Would remove ${paths.length} item(s) under ${result.target}`
-      : `Removed ${paths.length} item(s) under ${result.target}`,
-  ];
-  return joinSections(sections);
 }
 
 export function formatMutateResult(result: ResolveResult): string {
